+++ examples/shirt-shop/app/[code]/cart/page.tsx
@@ -1,5 +1,6 @@
-import { getCart } from '@/app/actions';
+import { getCart } from '@/utils/actions';
 import { CheckoutForm } from '@/components/checkout-form';
+import { Main } from '@/components/main';
 import { OrderSummary } from '@/components/order-summary';
 
 export default async function CartPage() {
@@ -9,11 +10,11 @@ export default async function CartPage() {
   const total = subtotal + shipping;
 
   return (
-    <main className="mx-auto max-w-2xl px-4 pb-16 sm:px-6 sm:pb-24 lg:max-w-7xl lg:px-8">
+    <Main>
       <div className="lg:grid lg:grid-cols-12 lg:items-start lg:gap-x-12 xl:gap-x-16">
         <OrderSummary items={items} />
         <CheckoutForm subtotal={subtotal} shipping={shipping} total={total} />
       </div>
-    </main>
+    </Main>
   );
 }
